requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==24.2.0
bidict==0.23.1
blinker==1.8.2
boto3==1.35.43
botocore==1.35.43
cachelib==0.13.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
deprecation==2.1.0
dnspython==2.7.0
eventlet==0.37.0
feedparser==6.0.11
Flask==3.0.3
Flask-Session==0.8.0
Flask-SocketIO==5.4.1
frozenlist==1.5.0
gotrue==2.11.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.1
msgspec==0.18.6
multidict==6.1.0
packaging==24.1
pefile==2024.8.26
postgrest==0.18.0
propcache==0.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
pywin32-ctypes==0.2.3
PyYAML==6.0.2
realtime==2.0.6
requests==2.32.3
s3transfer==0.10.3
setuptools==75.1.0
sgmllib3k==1.0.0
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
Werkzeug==3.0.4
wsproto==1.2.0
yarl==1.18.3
